requirements.txt
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
tqdm>=4.65.0 